Trying to improve the automatic elevation:
Allowing a maximum stream elevation - this is an attempt to cope with
streams at edges of the domain that may be sensibly incised by more than
the specififed minimum incision. - in the test example I have set this
to the 75th percentile value for surface elevations at streams -
arbitrary as hell.
Also trialing a prefered slope for segments which is based on the
top-surface slope between the segments ends.

This is all horribly arbitrary but at least the hooks exist - should
continue to refine as we play with more models.